rename folders (#9)
