+++ Sources/SwiftRulesEngine/ConditionEvaluator.swift
