+++ Sources/SwiftRulesEngine/ConditionRule.swift
