+++ Sources/SwiftRulesEngine/Evaluable.swift
