+++ Sources/SwiftRulesEngine/Evaluating.swift
