+++ Sources/SwiftRulesEngine/Expression/Operand.swift
