+++ Sources/SwiftRulesEngine/Operand+Literal.swift
