+++ Sources/SwiftRulesEngine/Result+RulesFailure.swift
