+++ Sources/SwiftRulesEngine/Template/MustacheToken.swift
