+++ Sources/SwiftRulesEngine/Template/Segment.swift
