+++ Sources/SwiftRulesEngine/Transforming.swift
