+++ Sources/SwiftRulesEngine/Traversable.swift
