+++ Tests/SwiftRulesEngineTests/swift-core/EventType.swift
