+++ Tests/SwiftRulesEngineTests/swift-core/SharedState.swift
